requirements.txt
Flask==3.0.0
requests==2.31.0
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
gunicorn==21.2.0
python-dotenv==1.0.0